Setup boilerplate, get tests setup
Add the transform() function to d3.svg, since this is an SVG-specfic
feature. Nothing but the name has been defined.

Get tests working by requiring d3 as a global variable. This makes it
available to the d3-transform module. There may be a better way to do
this, but I'm unaware of it.

Tests passing.