Add translate() to d3.svg.transform
translate() can be called in two ways:

  * with up to two arguments; or
  * with a function that will return an array of two elements.

In the first case, the first argument is used as dx and the second is
dy. If the second argument is omitted, then the translation is applied
to x, which is consistent with the spec.

If it's called with a function it's expected to return an array of up to
two arguments, or a single number. If the function returns an array,
they're treated as though they were positional arguments. If the
function returns a single number, it's treated as dx.

Also added associated tests.

Tests pass.