Add all other transformations to d3.svg.transform
All the SVG transformation types are now supported. No checks are in
place to ensure that the user provides the correct nubmer of arguments
to the transformations. This leaves the user in the same situation that
he was in befor the existence of this library.

Regardless, this code needs more tests.

Tests Pass